helia: processing reservation requests and traffic packets in dataplane
The border routers check all incoming packets for HBH extensions. If a reservation request is present,
its MAC is verified and reservation response is sent back. For traffic options, the BR checks is there is a reservation field for this AS and if so, increments the currRF field. The packet is forwarded and its helia parameters are printed to the logs.